move parsing logic to macro crate
This makes the backend crate solely concerned with having an ast
for which we can generate code.
